refactor: remove iternal for Archivista flexibility
Remove the internal in order to enable Archivista source code user to:
- Extend/Modify the existent services (config, storage, etc) within
  their code base

- Allow to use Archivista encapuslated in other services, example API
  service
