+++ pkg/metadatastorage/attestationcollection/parserstorer.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/in-toto/archivista/ent"
-	"github.com/in-toto/archivista/internal/metadatastorage"
+	"github.com/in-toto/archivista/pkg/metadatastorage"
 	"github.com/in-toto/go-witness/attestation"
 )
