+++ pkg/metadatastorage/sqlstore/store.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/digitorus/timestamp"
 	"github.com/in-toto/archivista/ent"
-	"github.com/in-toto/archivista/internal/metadatastorage"
-	"github.com/in-toto/archivista/internal/metadatastorage/parserregistry"
+	"github.com/in-toto/archivista/pkg/metadatastorage"
+	"github.com/in-toto/archivista/pkg/metadatastorage/parserregistry"
 	"github.com/in-toto/go-witness/cryptoutil"
 	"github.com/in-toto/go-witness/dsse"
 	"github.com/in-toto/go-witness/intoto"
