+++ pkg/server/server_test.go
@@ -27,9 +27,9 @@ import (
 	"testing"
 
 	"github.com/gorilla/mux"
-	"github.com/in-toto/archivista/internal/artifactstore"
-	"github.com/in-toto/archivista/internal/config"
 	"github.com/in-toto/archivista/pkg/api"
+	"github.com/in-toto/archivista/pkg/artifactstore"
+	"github.com/in-toto/archivista/pkg/config"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
 )
